bump: bumping version: 2.0.0b26.dev5 -> 2.0.0b26
BREAKING CHANGE: This version includes a framework class structure
refactor. Majority changes are made on how chat objects are structured.
Class names are changed to remove the “EFB” prefix. No operation logic
is changed.

Channels and middlewares updated before this version will most likely
NOT compatible with this version.

For developers, please see the summary below for detailed changes made
to the API.

https://www.notion.so/1a23/Framework-class-structure-refactor-summary-8eb7c610a3db4078afb4c94f122b6ecb